+++ app/Redirection.php
@@ -6,5 +6,8 @@ use Illuminate\Database\Eloquent\Model;
 
 class Redirection extends Model
 {
-    //
+    public function admin()
+    {
+        return $this->belongsTo('App\admin');
+    }
 }