+++ resources/assets/js/admin/views/donor/DetailDonorTransactionHistory.vue
@@ -158,6 +158,9 @@
   import Evidence from './DetailDonorTransactionHistoryEvidence'
   import Invoice from './DetailDonorTransactionHistoryInvoice'
   import flatPickr from 'vue-flatpickr-component';
+  import {
+    copyClipboard
+  } from '../_share/mixins/copyClipboard'
   import 'flatpickr/dist/flatpickr.css';
   import {
     FieldTableData
@@ -170,7 +173,7 @@
       Evidence,
       Invoice
     },
-    mixins: [FieldTableData],
+    mixins: [FieldTableData, copyClipboard],
     data: function() {
       return {
         confirmModal: false,
@@ -202,7 +205,6 @@
           invoice_no: '',
           status_invoice: null
         },
-        buttonCopyText: '<i class="fa fa-link"></i> Copy Link',
 
       }
     },
@@ -398,41 +400,6 @@
           status === 'NOT SENT' ? 'secondary' : 'primary'
       },
 
-      fallbackCopyTextToClipboard(text) {
-        var textArea = document.createElement("textarea");
-        textArea.value = text;
-
-        // Avoid scrolling to bottom
-        textArea.style.top = "0";
-        textArea.style.left = "0";
-        textArea.style.position = "fixed";
-
-        document.body.appendChild(textArea);
-        textArea.focus();
-        textArea.select();
-
-        try {
-          var successful = document.execCommand('copy');
-          var msg = successful ? 'successful' : 'unsuccessful';
-          console.log('Fallback: Copying text command was ' + msg);
-        } catch (err) {
-          console.error('Fallback: Oops, unable to copy', err);
-        }
-
-        document.body.removeChild(textArea);
-      },
-
-      copyTextToClipboard(text) {
-        if (!navigator.clipboard) {
-          this.fallbackCopyTextToClipboard(text);
-          return;
-        }
-        navigator.clipboard.writeText(text).then(function() {
-          console.log('Async: Copying to clipboard was successful!');
-        }, function(err) {
-          console.error('Async: Could not copy text: ', err);
-        });
-      },
       handleCopyClipboard() {
         let host = window.location.host
         let pathname = '/donor#/donation-confirmation?'